Massive rework on physics, networking, and loading.
This is quite a large commit since all the systems are quite interconnected with each other.

The entire login flow and the way the server starts up has been changed.  A package is no longer needed for a guest client; a future co-op mode will restore that functionality.  Single player and hosted modes avoid the normal login flow entirely.  The way the client handles the login has also changed, which includes new callbacks which the client uses to attach rendering components.  Furthermore, bugs with scene and object class creation on the client have been fixed.  Scene object collision data was given a new packet to better work with TMX scene objects.

TMX scene objects got a massive rewrite.  Instead of a scene object per chunk per layer, there is now a single scene object per layer.  Individual chunks of the map are loaded at-will and individual chunks are sent over to the client in a new packet type.  TMX scene objects can now be sent across the network to clients and don't require the map file to exist on the client PC.  Tilesets are also saved on the server and sent to clients.

The physics system obtained a new type of body: hybrid.  A hybrid body will be static when stationary and dynamic when moving.  This lets players body-block each other.  Scene object body configs are now stored so when a scene object swaps scenes, the physics can be recreated in the new scene's world.  The way physics get registered to bodies via Lua has also changed.  The user creates a body configuration and then assigns it to the scene object during the OnCreatePhysics callback.

The way scene objects get sent across the network has been improved.  The ability to turn off change replication should work correctly.  The owner of a scene object was changed to just be an id instead of a weak_ptr because the client may not yet know about the owning player.  Scene object ids have also been changed and a new class of scene object was created: local scene objects.  Local scene objects are not sent across the network, whereas a scene object with replication turned off will be sent across initially, but won't have updates sent.  The idea is that guests will be allowed to create local scene objects, but that is not yet implemented.  Adding, removing, and changing scene objects has been rewritten to be more correct across the network, and attribute update rates now work properly.

Finally, accounts have been a little improved.  The last known scene object is now saved on the account so players can reconnect to their appropriate scene.  Also, guest servers won't save accounts anymore.